backend/pyproject.toml
[tool.poetry]
name = "fastapi template"
version = "1.0.0"
description = "A scalable, optimized starter template with best practices for fast development and deployment."
authors = ["Yagnesh Vakharia <yagneshvakharia97@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yagnesh97/fastapi-template"
keywords = ["python", "fastapi", "template"]
package-mode = false

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/yagnesh97/fastapi-template/issues"

[tool.metadata]
title = "FastAPI Template"
full_description = """
FastAPI Template

📝 [Source Code](https://github.com/yagnesh97/fastapi-template)
🐞 [Issues](https://github.com/yagnesh97/fastapi-template/issues)
"""

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["standard"], version = "^0.115.5" }
uvicorn = { extras = ["standard"], version = "^0.32.1" }
pydantic-settings = "^2.6.1"
itsdangerous = "^2.2.0"
pyjwt = "^2.9.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pymongo = "^4.8.0"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
ruff = "^0.8.0"
pytest = "^8.3.3"
coverage = "^7.6.8"
httpx = "^0.27.2"
mongomock = "^4.3.0"
fakeredis = "^2.26.2"
types-redis = "^4.6.0.20241004"

[tool.ruff]
lint.select = ["E", "F", "I"]
lint.fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true
plugins = "pydantic.mypy"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
